pyproject.toml
[tool.poetry]
name = "apidev"
version = "0.1.0"
description = "FastAPI project to learn"
authors = ["Andreas Løfsgaard <andreas@lofsgaard.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "3.11.7"
fastapi = "0.109.2"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
python-multipart = "^0.0.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"
sqlmodel = "^0.0.14"
alembic = "^1.13.1"
httpx = "^0.26.0"
pytest = "^8.0.0"
bcrypt = "^4.1.2"
requests = "^2.31.0"
pytest-asyncio = "^0.23.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  ".", "app",
]
asyncio_mode = "auto"
